Update README.md
This code is used as a support of the article "On a three-dimensional tumor-virus compartmental model, and two four-dimensional oncolytic virotherapy models". It contains many computations used in the manuscript as well as some numerical simulations. 

To run some parts of the code, one needs to upload the EcoEvo package of Mathematica.